Refactor to es6
