+++ app.js
@@ -1,6 +1,7 @@
-const express = require('express');
-const logger = require('morgan');
-const bodyParser = require('body-parser');
+import express from 'express';
+import logger from 'morgan';
+import bodyParser from 'body-parser';
+import router from './server/routes';
 
 // Set up the express app
 const app = express();
@@ -12,7 +13,7 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-require('./server/routes')(app);
+router(app);
 
 // Setup a default catch-all route that sends back a 
 //welcome message in JSON format.
@@ -20,4 +21,4 @@ app.get('*', (req, res) => res.status(200).send({
   message: 'Welcome to the beginning of nothingness yo.',
 }));
 
-module.exports = app;
+export default app;
