+++ bin/www
@@ -1,7 +1,7 @@
 // This will be our application entry. We'll setup our server here.
 
-const http = require('http');
-const app = require('../app'); // The express app we just created
+import http from 'http';
+import app from '../app'; // The express app we just created
 
 const port = parseInt(process.env.PORT, 10) || 8080;
 app.set('port', port);
