+++ server/models/index.js
@@ -1,8 +1,6 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
+import fs from'fs';
+import path from 'path';
+import Sequelize from 'sequelize';
 
 const basename = path.basename(module.filename);
 const env = process.env.NODE_ENV || 'development';
@@ -39,4 +37,4 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-module.exports = db;
+export default db;
